Ignore subtitles when trying to detect black borders
In the current black border detection, any non-black pixel that's found
is concluded to be part of the image. As a result, if a letterboxed film
has subtitles in the bottom black area, the detected rectangle will
become larger. As a result, the LEDs will flash whenever the subtitles
appear or disappear, even though the scene itself isn't changing.

This commit changes the black border heuristics to ignore monochrome
pixels. Since almost all subtitles are white (or anti-aliased), and
almost all movies have color, this fixes the flashing for me.