Properly handle expired restore token on Wayland
The previous version just gets crashes on an exception from invalid
restore token and doesn't request a new valid one.

This version catches the exception and starts the screen cast with a new
token.

I'm not sure what the purpose of the 200 ms sleep is, so I left it
similar to the previous version.